style: Create source.go,sink.go files
